pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wedding-planning"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "pydantic-settings>=2.0.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "redis>=5.0.0",
    "jinja2>=3.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "D",  # pydocstyle
    "E",  # Error
    "F",  # pyflakes
    "I",  # isort
    "ISC", # flake8-implicit-str-concat
    "N",  # pep8-naming
    "PGH",  # pygrep-hooks
    "PTH", # flake8-use-pathlib
    "Q",  # flake8-quotes
    "S",  # bandit
    "SIM",  # flake8-simplify
    "TRY",  # tryceratops
    "UP",  # pyupgrade
    "W",  # Warning
    "YTT", # flake8-2020
]
ignore=[
    "D211",
    "D213",
    "D104",
    "D100",
    "D203",
    "D101",
    "D103",
    "E501"
]

[tool.ruff.format]
quote-style = "double"
